# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The problem of finding an optimal branch of VTB Bank",
        "title": "VTB MoreTech 5.0",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/api/atms": {
            "get": {
                "description": "Show all atms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "atm"
                ],
                "summary": "all atms",
                "operationId": "allAtms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.allAtmsResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
        "/api/offices": {
            "get": {
                "description": "Show all offices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "office"
                ],
                "summary": "all offices",
                "operationId": "allOffices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.allOfficesResponce"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.allAtmsResponce": {
            "type": "object"
        },
        "handler.allOfficesResponce": {
            "type": "object"
        },
        "handler.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}